refactor: create `adb` module, and add **LOTS** of docs
