+++ src/core/mod.rs
@@ -1,3 +1,4 @@
+pub mod adb;
 pub mod config;
 pub mod helpers;
 pub mod save;
